+++ flooding_worker/executor.py
@@ -1,11 +1,12 @@
 #!/usr/bin/python
 # (c) Nelen & Schuurmans.  GPL licensed.
 
-from flooding_worker.file_logging import setFileHandler, removeFileHandlers
-from flooding_worker.worker.action_workflow import ActionWorkflow, ActionTaskPublisher
+from flooding_worker.worker.action_workflow import (
+    ActionWorkflow, ActionTaskPublisher, ActionHeartbeat)
 from flooding_worker.worker.broker_connection import BrokerConnection
 from flooding_worker.worker.message_logging_handler import AMQPMessageHandler
-from flooding_worker.models import WorkflowTask, WorkflowTemplate
+from flooding_worker.models import WorkflowTask
+from django.conf import settings
 
 import logging
 log = logging.getLogger("flooding.management.start_scenario")
@@ -22,7 +23,7 @@ def start_workflow(scenario_id, workflowtemplate_id, log_level='INFO'):
     """
     numeric_level = getattr(logging, log_level.upper(), None)
     if not isinstance(numeric_level, int):
-        log.error("Invalid log level: %s" % options["log_level"])
+        log.error("Invalid log level: %s" % log_level)
         numeric_level = 10
 
     broker = BrokerConnection()
@@ -37,14 +38,15 @@ def start_workflow(scenario_id, workflowtemplate_id, log_level='INFO'):
     logging.handlers.AMQPMessageHandler = AMQPMessageHandler
     broker_handler = logging.handlers.AMQPMessageHandler(action,
                                                          numeric_level)
-    
+
     action.set_broker_logging_handler(broker_handler)
     status = action.perform_workflow()
 
     if connection.is_open:
         connection.close()
-    
+
     return status
+
 
 def start_task(task_id, log_level='INFO'):
     """
@@ -53,7 +55,7 @@ def start_task(task_id, log_level='INFO'):
     task = WorkflowTask.objects.get(pk=task_id)
     numeric_level = getattr(logging, log_level.upper(), None)
     if not isinstance(numeric_level, int):
-        log.error("Invalid log level: %s" % options["log_level"])
+        log.error("Invalid log level: %s" % log_level)
         numeric_level = 10
 
     broker = BrokerConnection()
@@ -67,7 +69,7 @@ def start_task(task_id, log_level='INFO'):
     logging.handlers.AMQPMessageHandler = AMQPMessageHandler
     broker_handler = logging.handlers.AMQPMessageHandler(action,
                                                          numeric_level)
-    
+
     action.set_broker_logging_handler(broker_handler)
     success = action.perform()
 
@@ -75,4 +77,30 @@ def start_task(task_id, log_level='INFO'):
         connection.close()
 
     return success
-    
+
+
+def start_heartbeat(log_level='INFO'):
+    """
+    Publish a message to execute a separate task.
+    """
+    numeric_level = getattr(logging, log_level.upper(), None)
+
+    broker = BrokerConnection()
+    connection = broker.connect_to_broker()
+    if connection is None:
+        log.error("Could not connect to broker.")
+        return
+
+    action = ActionHeartbeat(connection, settings.HEARTBEAT_QUEUES)
+
+    logging.handlers.AMQPMessageHandler = AMQPMessageHandler
+    broker_handler = logging.handlers.AMQPMessageHandler(action,
+                                                         numeric_level)
+
+    action.set_broker_logging_handler(broker_handler)
+    success = action.perform()
+
+    if connection.is_open:
+        connection.close()
+
+    return success